patch(vouched): created patch to vouchdao contract to support the registration of more than one service per wallet.
